models/diffusion/requirements.txt
torch
torchvision
diffusers>=0.30.0
PyYAML
pillow
matplotlib
tqdm
